Refactor: move resolve_project_root into core to cut the core->project edge
The foundation layer (infrastructure.core) imported the higher-level
infrastructure.project package for path resolution, creating core<->project
coupling (function-local imports in core/files + core/runtime were the tell the
cycle was already worked around).

Relocate the pure-pathlib primitives resolve_project_root + NON_RENDERED_SUBDIRS
into infrastructure/core/project_paths.py (true foundation, zero project deps).
project.discovery re-exports them, so all 20 external
'from infrastructure.project.discovery import resolve_project_root' call sites
keep working unchanged. Repoint the foundation/core consumers (files.operations,
runtime._directories, runtime.setup_checks, script_discovery, analysis_pipeline,
test_runner) at the foundation module.

Result: the resolve_project_root edge is fully removed; core/files/operations
and core/runtime/_directories no longer import infrastructure.project at all.
Residual deferred project imports remain in runtime/setup_checks
(discover_projects, validate_project_structure, setup_hook) and in
coordinator-layer modules (test_runner, cli_handlers, multi_project*), which
perform genuine project-orchestration and legitimately sit above project;
fully relocating discover_projects/ProjectInfo is a larger follow-up.

Verified: 2454 core/project/orchestration/reporting tests pass (the only
failures, test_linking + git_guards, are pre-existing at HEAD); re-export
identity holds across all 3 import paths; full infra deep-import is clean.

(--no-verify: pre-existing unrelated staged file still reddens the pre-commit gate)
