+++ infrastructure/core/runtime/setup_checks.py
@@ -52,7 +52,8 @@ def sync_workspace_dependencies(repo_root: Path) -> bool:
 
 
 def validate_project_discovery(repo_root: Path, project_name: str) -> bool:
-    from infrastructure.project.discovery import discover_projects, resolve_project_root
+    from infrastructure.core.project_paths import resolve_project_root
+    from infrastructure.project.discovery import discover_projects
     from infrastructure.project.validation import validate_project_structure
 
     logger.info("Discovering available projects...")
@@ -92,7 +93,7 @@ def validate_project_discovery(repo_root: Path, project_name: str) -> bool:
 
 
 def run_optional_setup_hook(repo_root: Path, project_name: str) -> bool:
-    from infrastructure.project.discovery import resolve_project_root
+    from infrastructure.core.project_paths import resolve_project_root
     from infrastructure.project.setup_hook import run_project_setup_hook
 
     project_dir = resolve_project_root(repo_root, project_name)
