+++ infrastructure/core/script_discovery.py
@@ -14,7 +14,7 @@ from typing import Any
 
 from infrastructure.core.exceptions import PipelineError
 from infrastructure.core.logging.utils import get_logger, log_success
-from infrastructure.project.discovery import resolve_project_root
+from infrastructure.core.project_paths import resolve_project_root
 
 logger = get_logger(__name__)
